requirements.txt
python-telegram-bot[job-queue]==20.3
requests>=2.31.0
python-dotenv>=1.0.0
Flask[async]>=2.3.0
pymongo
gunicorn